Clearer names and split the file into smaller bits
The renames are not-critical, because there has not been a release since
last merge.
